# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves the BHLQuest app. It locates relevant sections in the Biodiversity Heritage Library that correspond to a user's query. \\n\\nCode repository: https://github.com/gnames/bhlquest. \\n\\nAccess the API on the production server: https://bhlquest.globalnames.org/api/v1.",
        "title": "BHLQuest API",
        "contact": {
            "name": "Dmitry Mozzherin",
            "url": "https://github.com/dimus",
            "email": "dmozzherin@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "bhlquest.globalnames.org",
    "basePath": "/api/v1",
    "paths": {
...
        "/ask/{question}": {
            "get": {
                "description": "This endpoint receives a question about BHL and returns a list of pages containing the answer.",
                "produces": [
                    "application/json"
                ],
                "summary": "Ask a question",
                "operationId": "ask-question",
                "parameters": [
                    {
                        "type": "string",
                        "description": "A question to ask BHL about.",
                        "name": "question",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The maximum number or returned results.",
                        "name": "max-results",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "A score threshold from 0.0 to 1.0",
                        "name": "score-threshold",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Shows matched text in results",
                        "name": "with-text",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of pages containing the answer",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/output.Answer"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "output.Answer": {
            "description": "Answer holds the metadata and results for a query response, containing the pages from BHL that answer a given question.",
            "type": "object",
            "properties": {
                "metadata": {
                    "description": "Meta contains metadata about the request.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/output.Meta"
                        }
                    ]
                },
                "results": {
                    "description": "Results is a list of pages containing the answers.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/output.Result"
                    }
                },
                "summary": {
                    "description": "Summary contains an answer from LLM created from the\ncontent of the results.",
                    "type": "string"
                }
            }
        },
        "output.Meta": {
            "description": "Meta includes information such as the time taken to process the query.",
            "type": "object",
            "properties": {
                "maxResultsNum": {
                    "description": "MaxResultsNum is the maximum number of returned results.",
                    "type": "integer",
                    "example": 10
                },
                "queryTime": {
                    "description": "QueryTime is the duration taken to process the query.",
                    "type": "number",
                    "example": 0.911422974
                },
                "question": {
                    "description": "Question asked by user.",
                    "type": "string",
                    "example": "What are ecological niches for Indigo Bunting?"
                },
                "scoreThreshold": {
                    "description": "ScoreThreshold determines the smallest score which is\nstill considered for results.",
                    "type": "number",
                    "example": 0.4
                },
                "version": {
                    "description": "Version of BHLQuest",
                    "type": "string",
                    "example": "v0.0.3"
                }
            }
        },
        "output.Result": {
            "description": "Result holds information about a BHL page or range of pages that contain answers to a given question.",
            "type": "object",
            "properties": {
                "chunkId": {
                    "type": "integer",
                    "example": 2980234
                },
                "crossScore": {
                    "description": "CrossScore is generated by Cross-Embeding during comparison\nof the question with results. It is used for sorting results.",
                    "type": "number",
                    "example": 0.92353212
                },
                "itemId": {
                    "description": "ItemID is the ID of a BHL Item, such as a book\nor journal volume.",
                    "type": "integer",
                    "example": 226148
                },
                "language": {
                    "description": "Language is the main language of the item's title.",
                    "type": "string"
                },
                "outlink": {
                    "description": "Outlink is the URL pointing to the BHL web page\nfor PageID.",
                    "type": "string",
                    "example": "https://www.biodiversitylibrary.org/page/53469262"
                },
                "outlinkTitleDOI": {
                    "description": "OutlinkTitleDOI is the DOI of the item's title.",
                    "type": "string"
                },
                "pageId": {
                    "description": "PageID is the ID of the first page of the answer.",
                    "type": "integer",
                    "example": 53469262
                },
                "pageIndex": {
                    "description": "PageIndex is the index of the starting page of the answer.",
                    "type": "integer",
                    "example": 2
                },
                "pages": {
                    "description": "PageIDs is a list of IDs of pages in the Item.\nExample:\t[53469262, 53469263, 53469264, 53469265, 53469266]",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "reference": {
                    "description": "Reference is the string representation of the BHL reference.",
                    "type": "string"
                },
                "score": {
                    "description": "Score, generated by AI, indicates the relevance of\nthe result. Higher scores are better.",
                    "type": "number",
                    "example": 0.7505834773704542
                },
                "text": {
                    "description": "TextPages is the text from pages of the chunk.",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}